Added Swift playground for the demonstrated algorithms
For performance reasons, the actual implementations were moved to
.swift files (organized by category) to precompile them. Also, the
single editor would have cluttered up immensely.

Implementations generally stayed close to the JS-versions (also
including the original comments as beacons). Of course,  the function
names tend to be a bit “unswifty” as a result… not sure if that should
change.