Fix binary_search.js
## Why this was needed
While coding along with this section in the book, I was writing some tests to ensure I was doing everything right. The current version still hits an infinite loop when looking for an item that is in range of the list but no present in the list.

## How I fixed it
I noticed a simplification we could make in the "not found" state. Moving the min and max to exclude the current middle solves the problem I was having with the min and max getting stuck  at the same positions and removes the need for the explicit infinite loop check that was added.